Simplify/decomplect/reduce the get_all_server_details function
Switches the function to use Effects for its HTTP requests, instead
of directly using treq, and also get rid of a bunch of concerns
for things that shouldn't be concerns to this code, like retrying,
json decoding, URL configuration.

This is done by parameterizing the "request" function, and assuming
it has a very simple interface that implies all of those concerns.

This also lets us delete a bunch of test code!